+++ winter-android/src/test/kotlin/io/jentz/winter/android/AndroidPresentationScopeAdapterTest.kt
@@ -5,7 +5,6 @@ import android.app.Application
 import android.content.Context
 import android.content.ContextWrapper
 import android.view.View
-import com.nhaarman.mockito_kotlin.mock
 import com.nhaarman.mockito_kotlin.whenever
 import io.jentz.winter.GraphRegistry
 import io.jentz.winter.WinterException
@@ -15,8 +14,6 @@ import io.kotlintest.matchers.boolean.shouldBeTrue
 import io.kotlintest.matchers.types.shouldBeSameInstanceAs
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldThrow
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertSame
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -24,7 +21,7 @@ import org.mockito.Mock
 import org.mockito.junit.MockitoJUnitRunner
 
 @RunWith(MockitoJUnitRunner::class)
-class SimpleAndroidInjectionAdapterTest {
+class AndroidPresentationScopeAdapterTest {
 
     @Mock private lateinit var application: Application
     @Mock private lateinit var context: Context
@@ -32,11 +29,12 @@ class SimpleAndroidInjectionAdapterTest {
     @Mock private lateinit var view: View
     @Mock private lateinit var contextWrapper: ContextWrapper
 
-    private val adapter = SimpleAndroidInjectionAdapter()
+    private val adapter = AndroidPresentationScopeAdapter()
 
     private val applicationComponent = component {
-        constant("application")
-        subcomponent("activity") {
+        subcomponent("presentation") {
+            subcomponent("activity") {
+            }
         }
     }
 
@@ -50,7 +48,7 @@ class SimpleAndroidInjectionAdapterTest {
     fun `#createGraph with an Application instance should open root graph with application as constant`() {
         val graph = adapter.createGraph(application, null)
 
-        graph.instance<String>().shouldBe("application")
+        graph.component.qualifier.shouldBe(null)
         graph.instance<Application>().shouldBe(application)
         graph.instance<Context>().shouldBe(application)
     }
@@ -62,18 +60,30 @@ class SimpleAndroidInjectionAdapterTest {
             constant(instance)
         }
 
-        graph.instance<String>().shouldBe("application")
+        graph.component.qualifier.shouldBe(null)
         graph.instance<Any>().shouldBeSameInstanceAs(instance)
     }
 
     @Test
-    fun `#createGraph with an Activity instance should open activity graph with activity as constant`() {
+    fun `#createGraph with an Activity instance should open presentation and activity graph with activity as constant`() {
         GraphRegistry.open()
         val graph = adapter.createGraph(activity, null)
 
         graph.component.qualifier.shouldBe("activity")
         graph.instance<Activity>().shouldBe(activity)
         graph.instance<Context>().shouldBe(activity)
+
+        graph.parent?.component?.qualifier.shouldBe("presentation")
+    }
+
+    @Test
+    fun `#createGraph with an Activity instance and a builder block should apply that builder block to component init`() {
+        GraphRegistry.open()
+        val instance = Any()
+        val graph = adapter.createGraph(activity) {
+            constant(instance)
+        }
+        graph.instance<Any>().shouldBeSameInstanceAs(instance)
     }
 
     @Test
@@ -92,8 +102,11 @@ class SimpleAndroidInjectionAdapterTest {
 
     @Test
     fun `#getGraph called with activity should get graph from GraphRegistry`() {
+        val presentationIdentifier = activity.javaClass.name
         GraphRegistry.open()
-        val graph = GraphRegistry.open("activity", identifier = activity)
+        GraphRegistry.open("presentation", identifier = presentationIdentifier)
+
+        val graph = GraphRegistry.open(presentationIdentifier, "activity", identifier = activity)
         adapter.getGraph(activity).shouldBe(graph)
     }
 
@@ -123,15 +136,36 @@ class SimpleAndroidInjectionAdapterTest {
     fun `#disposeGraph with Application instance should close root graph`() {
         GraphRegistry.open()
         adapter.disposeGraph(application)
-        GraphRegistry.has().shouldBe(false)
+        GraphRegistry.has().shouldBeFalse()
     }
 
     @Test
-    fun `#disposeGraph with Activity instance should close Activity graph`() {
+    fun `#disposeGraph with Activity instance should only close Activity graph`() {
+        val presentationIdentifier = activity.javaClass.name
         GraphRegistry.open()
-        GraphRegistry.open("activity", identifier = activity)
+        GraphRegistry.open("presentation", identifier = presentationIdentifier)
+        GraphRegistry.open(presentationIdentifier, "activity", identifier = activity)
+
+        GraphRegistry.has(presentationIdentifier).shouldBeTrue()
+        GraphRegistry.has(presentationIdentifier, activity).shouldBeTrue()
         adapter.disposeGraph(activity)
-        GraphRegistry.has(activity).shouldBeFalse()
+        GraphRegistry.has(presentationIdentifier).shouldBeTrue()
+        GraphRegistry.has(presentationIdentifier, activity).shouldBeFalse()
     }
+
+    @Test
+    fun `#disposeGraph with Activity instance should close Activity graph when Activity is finishing`() {
+        val presentationIdentifier = activity.javaClass.name
+        GraphRegistry.open()
+        GraphRegistry.open("presentation", identifier = presentationIdentifier)
+        GraphRegistry.open(presentationIdentifier, "activity", identifier = activity)
+
+        whenever(activity.isFinishing).thenReturn(true)
+
+        GraphRegistry.has(presentationIdentifier).shouldBeTrue()
+        adapter.disposeGraph(activity)
+        GraphRegistry.has(presentationIdentifier).shouldBeFalse()
+    }
+
 
 }