+++ winter/src/main/kotlin/io/jentz/winter/Injection.kt
@@ -75,11 +75,12 @@ object Injection {
          * The adapter implementation is responsible for storing the created graph.
          *
          * @param instance The instance to create a dependency graph for.
+         * @param builderBlock An optional builder block to pass to the component init method.
          * @return The newly created graph
          * @throws [io.jentz.winter.WinterException] if given [instance] type is not supported.
          *
          */
-        fun createGraph(instance: Any): Graph
+        fun createGraph(instance: Any, builderBlock: ComponentBuilderBlock?): Graph
 
         /**
          * Dispose the dependency graph of the given [instance].
@@ -98,7 +99,10 @@ object Injection {
     class ApplicationGraphOnlyAdapter : Adapter {
         override fun getGraph(instance: Any): Graph = GraphRegistry.get()
 
-        override fun createGraph(instance: Any): Graph = GraphRegistry.create()
+        override fun createGraph(
+            instance: Any,
+            builderBlock: ComponentBuilderBlock?
+        ): Graph = GraphRegistry.create(builderBlock = builderBlock)
 
         override fun disposeGraph(instance: Any) {
             GraphRegistry.close()
@@ -116,11 +120,13 @@ object Injection {
      * Create and return dependency graph for [instance].
      *
      * @param instance The instance for which a graph should be created.
+     * @param builderBlock An optional builder block to pass to the component init method.
      * @return The newly created graph.
      * @throws [io.jentz.winter.WinterException] if given [instance] type is not supported.
      */
     @JvmStatic
-    fun createGraph(instance: Any): Graph = adapter.createGraph(instance)
+    fun createGraph(instance: Any, builderBlock: ComponentBuilderBlock? = null): Graph =
+        adapter.createGraph(instance, builderBlock)
 
     /**
      * Create and return dependency graph for [instance] and also pass the graph to the given
@@ -128,12 +134,16 @@ object Injection {
      *
      * @param instance The instance for which a graph should be created.
      * @param injector The injector to inject into.
+     * @param builderBlock An optional builder block to pass to the component init method.
      * @return The created dependency graph.
      * @throws [io.jentz.winter.WinterException] if given [instance] type is not supported.
      */
     @JvmStatic
-    fun createGraphAndInject(instance: Any, injector: Injector): Graph =
-        createGraph(instance).also(injector::inject)
+    fun createGraphAndInject(
+        instance: Any,
+        injector: Injector,
+        builderBlock: ComponentBuilderBlock? = null
+    ): Graph = createGraph(instance, builderBlock).also(injector::inject)
 
     /**
      * Create and return dependency graph for [instance] and inject all members into instance.
@@ -142,13 +152,19 @@ object Injection {
      *
      * @param instance The instance to create a graph for and to inject into.
      * @param injectSuperClasses If true this will look for members injectors for super classes too.
+     * @param builderBlock An optional builder block to pass to the component init method.
      * @return The created dependency graph.
      * @throws [io.jentz.winter.WinterException] if given [instance] type is not supported.
      */
     @JvmStatic
     @JvmOverloads
-    fun <T : Any> createGraphAndInject(instance: T, injectSuperClasses: Boolean = false): Graph =
-        createGraph(instance).also { graph -> graph.inject(instance, injectSuperClasses) }
+    fun <T : Any> createGraphAndInject(
+        instance: T,
+        injectSuperClasses: Boolean = false,
+        builderBlock: ComponentBuilderBlock? = null
+    ): Graph = createGraph(instance, builderBlock).also { graph ->
+        graph.inject(instance, injectSuperClasses)
+    }
 
     /**
      * Get dependency graph for [instance].